requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-docx
python-dotenv
Pillow
PyMuPDF 
transformers 
torch 
huggingface_hub 
accelerate 